Azure Notebooks problem 4 markdown relative image
Attempt workarounds - previous relative images used to work in
Azure NB but are now broken. So try replacing with links to GitHub
versions or in the case of the svg by direct inclusion (ugly).

Notebook looks fine locally but will it work with Azure?

Other people have reported this problem

https://github.com/Microsoft/AzureNotebooks/issues/534